// projects/bible/web/src/interfaces/ISongsBook.ts
// Generated by https://quicktype.io

export interface ISongsBookResponses {
  status: string;
  results: number;
  data: Data;
}

export interface Data {
  data: ISong[];
}

export interface ISong {
  _id: string;
  book: Book;
  views: number;
  likes: number;
  num: number;
  title: string;
  lyrics: string;
  videos: any[];
  language: Language;
  songId: string;
  lyricsMarkdown: string;
  lyricsHtml: string;
  slug: string;
}

export enum Book {
  The63C7077F7E0B612Ee4C7E96A = '63c7077f7e0b612ee4c7e96a',
}

export enum Language {
  Fr = 'fr',
  HT = 'ht',
}